+++ objectbox/lib/src/native/sync.dart
@@ -175,8 +175,7 @@ class SyncChange {
   SyncChange._(this.entityId, this.entity, this.puts, this.removals);
 }
 
-/// Sync client is used to connect to an ObjectBox sync server.
-/// Create a client with [Sync.client] and related methods.
+/// A Sync client is used to connect to an ObjectBox Sync server.
 class SyncClient {
   final Store _store;
 
@@ -189,16 +188,47 @@ class SyncClient {
       : throw StateError('SyncClient already closed');
 
   /// Creates a Sync client associated with the given store and options.
+  /// This does not initiate any connection attempts yet, call [start] to do so.
   ///
-  /// This is an internal constructor. Clients should be created using
-  /// [Sync.client] and related helper methods.
-  SyncClient._(
-      this._store,
-      List<String> serverUrls,
-      List<SyncCredentials> credentials,
-      Map<String, String>? filterVariables,
+  /// By default, a Sync client automatically receives updates from the server
+  /// once login succeeded. To configure this differently, call
+  /// [setRequestUpdatesMode] with the wanted mode.
+  ///
+  /// Typically, only a single URL and a single credentials object is passed:
+  ///
+  /// ```dart
+  /// // Connect to a test server running on localhost using
+  /// // an unencrypted connection without authentication
+  /// SyncClient client =
+  ///     SyncClient(store, ['ws://127.0.0.1:9999'], [SyncCredentials.none()]);
+  /// ```
+  ///
+  /// Passing multiple URLs allows high availability and load balancing (for ex.
+  /// using a ObjectBox Sync Server Cluster). A random URL is selected for each
+  /// connection attempt.
+  ///
+  /// When passing multiple credentials, can't include [SyncCredentials.none].
+  ///
+  /// To configure [Sync filter](https://sync.objectbox.io/sync-server/sync-filters)
+  /// variables, pass variable names mapped to their value to [filterVariables].
+  ///
+  /// Sync client filter variables can be used in server-side Sync filters to
+  /// filter out objects that do not match the filter.
+  ///
+  /// To, for example, use self-signed certificates in a local development
+  /// environment or custom CAs, pass certificate paths referring to the local
+  /// file system to [certificatePaths].
+  ///
+  /// To configure Sync behavior, pass bitwise OR-ed [OBXSyncFlags] values to
+  /// [flags]. See [OBXSyncFlags] for available flags.
+  SyncClient(
+      this._store, List<String> serverUrls, List<SyncCredentials> credentials,
+      {Map<String, String>? filterVariables,
       List<String>? certificatePaths,
-      int? flags) {
+      int? flags}) {
+    if (syncClientsStorage.containsKey(_store)) {
+      throw StateError('Only one sync client can be active for a store');
+    }
     if (serverUrls.isEmpty) {
       throw ArgumentError.value(
           serverUrls, "serverUrls", "Provide at least one server URL");
@@ -245,6 +275,9 @@ class SyncClient {
       // also covers the length == 0 case
       setMultipleCredentials(credentials);
     }
+
+    syncClientsStorage[_store] = this;
+    InternalStoreAccess.addCloseListener(_store, this, close);
   }
 
   /// Closes and cleans up all resources used by this sync client.
@@ -747,7 +780,7 @@ class _SyncListenerGroup<StreamValueType> {
 /// [ObjectBox Sync](https://objectbox.io/sync/) makes data available and
 /// synchronized across devices, online and offline.
 ///
-/// Create a client using [Sync.client].
+/// Create a client using the [SyncClient] constructor.
 class Sync {
   /// Set to `true` to enable shared global IDs for a Sync-enabled entity class.
   ///
@@ -793,12 +826,13 @@ class Sync {
   ///
   /// To configure Sync behavior, pass bitwise OR-ed [OBXSyncFlags] values to
   /// [flags]. See [OBXSyncFlags] for available flags.
+  @Deprecated('Use the SyncClient constructor instead')
   static SyncClient client(
           Store store, String serverUrl, SyncCredentials credentials,
           {Map<String, String>? filterVariables,
           List<String>? certificatePaths,
           int? flags}) =>
-      clientMultiUrls(store, [serverUrl], credentials,
+      SyncClient(store, [serverUrl], [credentials],
           filterVariables: filterVariables,
           certificatePaths: certificatePaths,
           flags: flags);
@@ -807,12 +841,13 @@ class Sync {
   ///
   /// When passing multiple credentials, does **not** support
   /// [SyncCredentials.none].
+  @Deprecated('Use the SyncClient constructor instead')
   static SyncClient clientMultiCredentials(
           Store store, String serverUrl, List<SyncCredentials> credentials,
           {Map<String, String>? filterVariables,
           List<String>? certificatePaths,
           int? flags}) =>
-      clientMultiCredentialsMultiUrls(store, [serverUrl], credentials,
+      SyncClient(store, [serverUrl], credentials,
           filterVariables: filterVariables,
           certificatePaths: certificatePaths,
           flags: flags);
@@ -822,12 +857,13 @@ class Sync {
   /// Passing multiple URLs allows high availability and load balancing (for ex.
   /// using a ObjectBox Sync Server Cluster). A random URL is selected for each
   /// connection attempt.
+  @Deprecated('Use the SyncClient constructor instead')
   static SyncClient clientMultiUrls(
           Store store, List<String> serverUrls, SyncCredentials credentials,
           {Map<String, String>? filterVariables,
           List<String>? certificatePaths,
           int? flags}) =>
-      clientMultiCredentialsMultiUrls(store, serverUrls, [credentials],
+      SyncClient(store, serverUrls, [credentials],
           filterVariables: filterVariables,
           certificatePaths: certificatePaths,
           flags: flags);
@@ -837,18 +873,14 @@ class Sync {
   ///
   /// When passing multiple credentials, does **not** support
   /// [SyncCredentials.none].
-  static SyncClient clientMultiCredentialsMultiUrls(
-      Store store, List<String> serverUrls, List<SyncCredentials> credentials,
-      {Map<String, String>? filterVariables,
-      List<String>? certificatePaths,
-      int? flags}) {
-    if (syncClientsStorage.containsKey(store)) {
-      throw StateError('Only one sync client can be active for a store');
-    }
-    final client = SyncClient._(store, serverUrls, credentials, filterVariables,
-        certificatePaths, flags);
-    syncClientsStorage[store] = client;
-    InternalStoreAccess.addCloseListener(store, client, client.close);
-    return client;
-  }
+  @Deprecated('Use the SyncClient constructor instead')
+  static SyncClient clientMultiCredentialsMultiUrls(Store store,
+          List<String> serverUrls, List<SyncCredentials> credentials,
+          {Map<String, String>? filterVariables,
+          List<String>? certificatePaths,
+          int? flags}) =>
+      SyncClient(store, serverUrls, credentials,
+          filterVariables: filterVariables,
+          certificatePaths: certificatePaths,
+          flags: flags);
 }